keep original 1wallets.js, spin out separate file
